+++ thrust/system/hip/hipstdpar/impl/search.hpp
@@ -107,6 +107,152 @@ find_first_of(InputIt1 first, InputIt1 last,
 
     return first + offset;
 }
+
+// BEGIN SEARCH
+template<class InputIt1, class InputIt2, class BinaryPred>
+InputIt1 THRUST_HIP_FUNCTION
+search(InputIt1 first, InputIt1 last,
+              InputIt2 s_first, InputIt2 s_last,
+              BinaryPred p)
+{
+    if (s_first == s_last)
+    {
+        return first;
+    }
+
+    thrust::device_system_tag dev_tag;
+    size_t* d_output;
+    d_output = thrust::malloc<size_t>(dev_tag, sizeof(*d_output)).get();
+
+    size_t size = last - first;
+    size_t s_size = s_last - s_first;
+
+    // temp storage
+    size_t temp_storage_size_bytes;
+    void*  d_temp_storage = nullptr;
+    // Get size of d_temp_storage
+    hipError_t error = ::rocprim::search(d_temp_storage,
+                                        temp_storage_size_bytes,
+                                        first,
+                                        s_first,
+                                        d_output,
+                                        size,
+                                        s_size,
+                                        p);
+
+    if (error != hipSuccess)
+    {
+        return last;
+    }
+
+    d_temp_storage = thrust::malloc(dev_tag, temp_storage_size_bytes).get();
+
+    error = ::rocprim::search(d_temp_storage,
+                            temp_storage_size_bytes,
+                            first,
+                            s_first,
+                            d_output,
+                            size,
+                            s_size,
+                            p);
+
+    if (error != hipSuccess)
+    {
+        thrust::free(dev_tag, d_temp_storage);
+        thrust::free(dev_tag, d_output);
+        return last;
+    }
+
+    error = hipDeviceSynchronize();
+    if (error != hipSuccess)
+    {
+        thrust::free(dev_tag, d_temp_storage);
+        thrust::free(dev_tag, d_output);
+        return last;
+    }
+
+    size_t offset;
+    hipMemcpy(&offset, d_output, sizeof(offset), hipMemcpyDeviceToHost);
+
+    thrust::free(dev_tag, d_temp_storage);
+    thrust::free(dev_tag, d_output);
+
+    return first + offset;
+}
+// END SEARCH
+
+// BEGIN FIND_END
+template<class InputIt1, class InputIt2, class BinaryPred>
+InputIt1 THRUST_HIP_FUNCTION
+find_end(InputIt1 first, InputIt1 last,
+              InputIt2 s_first, InputIt2 s_last,
+              BinaryPred p)
+{
+    if (s_first == s_last)
+    {
+        return last;
+    }
+
+    thrust::device_system_tag dev_tag;
+    size_t* d_output;
+    d_output = thrust::malloc<size_t>(dev_tag, sizeof(*d_output)).get();
+
+    size_t size = last - first;
+    size_t s_size = s_last - s_first;
+
+    // temp storage
+    size_t temp_storage_size_bytes;
+    void*  d_temp_storage = nullptr;
+    // Get size of d_temp_storage
+    hipError_t error = ::rocprim::find_end(d_temp_storage,
+                                        temp_storage_size_bytes,
+                                        first,
+                                        s_first,
+                                        d_output,
+                                        size,
+                                        s_size,
+                                        p);
+
+    if (error != hipSuccess)
+    {
+        return last;
+    }
+
+    d_temp_storage = thrust::malloc(dev_tag, temp_storage_size_bytes).get();
+
+    error = ::rocprim::find_end(d_temp_storage,
+                            temp_storage_size_bytes,
+                            first,
+                            s_first,
+                            d_output,
+                            size,
+                            s_size,
+                            p);
+
+    if (error != hipSuccess)
+    {
+        thrust::free(dev_tag, d_temp_storage);
+        thrust::free(dev_tag, d_output);
+        return last;
+    }
+
+    error = hipDeviceSynchronize();
+    if (error != hipSuccess)
+    {
+        thrust::free(dev_tag, d_temp_storage);
+        thrust::free(dev_tag, d_output);
+        return last;
+    }
+
+    size_t offset;
+    hipMemcpy(&offset, d_output, sizeof(offset), hipMemcpyDeviceToHost);
+
+    thrust::free(dev_tag, d_temp_storage);
+    thrust::free(dev_tag, d_output);
+
+    return first + offset;
+}
+// END FIND_END
 }
 
 namespace std
@@ -309,7 +455,61 @@ namespace std
     // END FIND_IF_NOT
 
     // BEGIN FIND_END
-    // TODO: UNIMPLEMENTED IN THRUST
+        template< class ForwardIt1,
+              class ForwardIt2,
+              enable_if_t<
+                ::hipstd::is_offloadable_iterator<ForwardIt1>() &&
+                ::hipstd::is_offloadable_iterator<ForwardIt2>()>* = nullptr>
+        inline
+        ForwardIt1 find_end(execution::parallel_unsequenced_policy, 
+                                 ForwardIt1 first, ForwardIt1 last,
+                                 ForwardIt2 s_first, ForwardIt2 s_last)
+    {
+        return ::thrust::find_end(first, last, s_first, s_last, thrust::equal_to<> {});
+    }
+
+    template< class ForwardIt1,
+              class ForwardIt2,
+              enable_if_t<
+                !::hipstd::is_offloadable_iterator<ForwardIt1>() ||
+                !::hipstd::is_offloadable_iterator<ForwardIt2>()>* = nullptr>
+        inline
+        ForwardIt1 find_end(execution::parallel_unsequenced_policy, 
+                                 ForwardIt1 first, ForwardIt1 last,
+                                 ForwardIt2 s_first, ForwardIt2 s_last )
+    {
+        return ::std::find_end(::std::execution::par, first, last, s_first, s_last);
+    }
+
+    template< class ForwardIt1,
+              class ForwardIt2,
+              class BinaryPred,
+              enable_if_t<
+                ::hipstd::is_offloadable_iterator<ForwardIt1>() &&
+                ::hipstd::is_offloadable_iterator<ForwardIt2>() &&
+                ::hipstd::is_offloadable_callable<BinaryPred>()>* = nullptr>
+        inline
+        ForwardIt1 find_end(execution::parallel_unsequenced_policy, 
+                                 ForwardIt1 first, ForwardIt1 last,
+                                 ForwardIt2 s_first, ForwardIt2 s_last, BinaryPred p)
+    {
+        return ::thrust::find_end(first, last, s_first, s_last, p);
+    }
+
+    template< class ForwardIt1,
+              class ForwardIt2,
+              class BinaryPred,
+              enable_if_t<
+                !::hipstd::is_offloadable_iterator<ForwardIt1>() ||
+                !::hipstd::is_offloadable_iterator<ForwardIt2>() ||
+                !::hipstd::is_offloadable_callable<BinaryPred>()>* = nullptr>
+        inline
+        ForwardIt1 find_end(execution::parallel_unsequenced_policy, 
+                                 ForwardIt1 first, ForwardIt1 last,
+                                 ForwardIt2 s_first, ForwardIt2 s_last, BinaryPred p)
+    {
+        return ::std::find_end(::std::execution::par, first, last, s_first, s_last, p);
+    }
     // END FIND_END
 
     // BEGIN FIND_FIRST_OF
@@ -803,7 +1003,61 @@ namespace std
     // END EQUAL
 
     // BEGIN SEARCH
-    // TODO: UNIMPLEMENTED IN THRUST
+    template< class ForwardIt1,
+              class ForwardIt2,
+              enable_if_t<
+                ::hipstd::is_offloadable_iterator<ForwardIt1>() &&
+                ::hipstd::is_offloadable_iterator<ForwardIt2>()>* = nullptr>
+        inline
+        ForwardIt1 search(execution::parallel_unsequenced_policy, 
+                                 ForwardIt1 first, ForwardIt1 last,
+                                 ForwardIt2 s_first, ForwardIt2 s_last)
+    {
+        return ::thrust::search(first, last, s_first, s_last, thrust::equal_to<> {});
+    }
+
+    template< class ForwardIt1,
+              class ForwardIt2,
+              enable_if_t<
+                !::hipstd::is_offloadable_iterator<ForwardIt1>() ||
+                !::hipstd::is_offloadable_iterator<ForwardIt2>()>* = nullptr>
+        inline
+        ForwardIt1 search(execution::parallel_unsequenced_policy, 
+                                 ForwardIt1 first, ForwardIt1 last,
+                                 ForwardIt2 s_first, ForwardIt2 s_last )
+    {
+        return ::std::search(::std::execution::par, first, last, s_first, s_last);
+    }
+
+    template< class ForwardIt1,
+              class ForwardIt2,
+              class BinaryPred,
+              enable_if_t<
+                ::hipstd::is_offloadable_iterator<ForwardIt1>() &&
+                ::hipstd::is_offloadable_iterator<ForwardIt2>() &&
+                ::hipstd::is_offloadable_callable<BinaryPred>()>* = nullptr>
+        inline
+        ForwardIt1 search(execution::parallel_unsequenced_policy, 
+                                 ForwardIt1 first, ForwardIt1 last,
+                                 ForwardIt2 s_first, ForwardIt2 s_last, BinaryPred p)
+    {
+        return ::thrust::search(first, last, s_first, s_last, p);
+    }
+
+    template< class ForwardIt1,
+              class ForwardIt2,
+              class BinaryPred,
+              enable_if_t<
+                !::hipstd::is_offloadable_iterator<ForwardIt1>() ||
+                !::hipstd::is_offloadable_iterator<ForwardIt2>() ||
+                !::hipstd::is_offloadable_callable<BinaryPred>()>* = nullptr>
+        inline
+        ForwardIt1 search(execution::parallel_unsequenced_policy, 
+                                 ForwardIt1 first, ForwardIt1 last,
+                                 ForwardIt2 s_first, ForwardIt2 s_last, BinaryPred p)
+    {
+        return ::std::search(::std::execution::par, first, last, s_first, s_last, p);
+    }
     // END SEARCH
 
     // BEGIN SEARCH_N